Move files to corresponding projects
