+++ AgGridPgsql.sln
@@ -5,6 +5,10 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api", "Api\Api.csproj", "{32320255-5980-4E9E-B214-C642C122AF9B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QueryBuilder", "QueryBuilder\QueryBuilder.csproj", "{A867FEAC-9690-4777-83E0-9263D33C6B4C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Contracts", "Contracts\Contracts.csproj", "{53FDDCA1-F243-4C92-A8FC-56D8E5D8BBA8}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -18,5 +22,13 @@ Global
 		{32320255-5980-4E9E-B214-C642C122AF9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{32320255-5980-4E9E-B214-C642C122AF9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{32320255-5980-4E9E-B214-C642C122AF9B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A867FEAC-9690-4777-83E0-9263D33C6B4C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A867FEAC-9690-4777-83E0-9263D33C6B4C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A867FEAC-9690-4777-83E0-9263D33C6B4C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A867FEAC-9690-4777-83E0-9263D33C6B4C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{53FDDCA1-F243-4C92-A8FC-56D8E5D8BBA8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{53FDDCA1-F243-4C92-A8FC-56D8E5D8BBA8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{53FDDCA1-F243-4C92-A8FC-56D8E5D8BBA8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{53FDDCA1-F243-4C92-A8FC-56D8E5D8BBA8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
