+++ Api/Api.csproj
@@ -6,12 +6,10 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <UserSecretsId>b8a07d7e-4190-425c-a13b-f44b9c4f84a7</UserSecretsId>
   </PropertyGroup>
-
+  
   <ItemGroup>
-    <PackageReference Include="Npgsql" Version="7.0.1" />
-    <PackageReference Include="Npgsql.DependencyInjection" Version="7.0.1" />
-    <PackageReference Include="SqlKata" Version="2.4.0" />
-    <PackageReference Include="SqlKata.Execution" Version="2.4.0" />
+    <ProjectReference Include="..\Contracts\Contracts.csproj" />
+    <ProjectReference Include="..\QueryBuilder\QueryBuilder.csproj" />
   </ItemGroup>
 
 </Project>
